Added FRAPP API
Extended Category model with color and description.
Extended Note model with field audio_url, to be able to playback audio files directly in the mobile app.
The field is hidden and will only be used to automatically retrieve the direct file's URL from CBA if the settings variable CBA_API_KEY is set and valid.

See #24